It's not safe to assume encrypted_password is always NOT NULL, so don't try to save old password if it's nil.

+++ lib/devise-security/models/password_archivable.rb
@@ -58,7 +58,7 @@ module Devise
 
       # archive the last password before save and delete all to old passwords from archive
       def archive_password
-        if encrypted_password_changed?
+        if encrypted_password_changed? && encrypted_password_change.first.present?
           if archive_count.to_i > 0
             old_passwords.create! old_password_params
             old_passwords.order(:id).reverse_order.offset(archive_count).destroy_all
